# tiendita-v6/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Tiendita API
  version: 1.0.0
  description: Documentación combinada de los módulos de la API
paths:
...
  /auth/register:
    post:
      summary: Registrar nuevo usuario
      description: Crea un nuevo usuario con rol `ROLE_USER` por defecto.
      tags:
      - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado correctamente (sin token)
        '400':
          description: Email ya registrado o datos inválidos
  /auth/login:
    post:
      summary: Autenticación de usuario
      description: Valida credenciales y devuelve un JWT con roles asociados.
      tags:
      - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso, devuelve token y datos de usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciales inválidas
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      required:
      - nombre
      - email
      - password
      properties:
        nombre:
          type: string
          example: Juan Pérez
        email:
          type: string
          format: email
          example: juan@example.com
        password:
          type: string
          format: password
          example: secret123
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
          example: juan@example.com
        password:
          type: string
          format: password
          example: secret123
    LoginResponse:
      type: object
      properties:
        nombre:
          type: string
          example: Juan Pérez
        email:
          type: string
          format: email
          example: juan@example.com
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        roles:
          type: array
          items:
            type: string
          example:
          - ROLE_USER